# backEnd/src/SeattleByNight.Api/openapi/SeattleByNight.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SeattleByNight.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/world/rooms/{roomId}": {
      "get": {
        "tags": [
          "WorldEditorEndpoints"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "WorldEditorEndpoints"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorldRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WorldEditorEndpoints"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "relocateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/world/rooms": {
      "post": {
        "tags": [
          "WorldEditorEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorldRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/world/rooms/{roomId}/deletable": {
      "get": {
        "tags": [
          "WorldEditorEndpoints"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/world/exits": {
      "post": {
        "tags": [
          "WorldEditorEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorldExitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/world/exits/{exitId}": {
      "put": {
        "tags": [
          "WorldEditorEndpoints"
        ],
        "parameters": [
          {
            "name": "exitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorldExitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorldExitRequest": {
        "required": [
          "sourceRoomId",
          "destinationRoomId",
          "direction",
          "isHidden",
          "isLocked"
        ],
        "type": "object",
        "properties": {
          "sourceRoomId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationRoomId": {
            "type": "string",
            "format": "uuid"
          },
          "direction": {
            "type": [
              "null",
              "string"
            ]
          },
          "isHidden": {
            "type": "boolean"
          },
          "isLocked": {
            "type": "boolean"
          }
        }
      },
      "CreateWorldRoomRequest": {
        "required": [
          "name",
          "description",
          "accessType",
          "mapX",
          "mapY",
          "mapLayer"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "accessType": {
            "type": [
              "null",
              "string"
            ]
          },
          "mapX": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "mapY": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "mapLayer": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "UpdateWorldExitRequest": {
        "required": [
          "sourceRoomId",
          "destinationRoomId",
          "direction",
          "isHidden",
          "isLocked",
          "version"
        ],
        "type": "object",
        "properties": {
          "sourceRoomId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationRoomId": {
            "type": "string",
            "format": "uuid"
          },
          "direction": {
            "type": [
              "null",
              "string"
            ]
          },
          "isHidden": {
            "type": "boolean"
          },
          "isLocked": {
            "type": "boolean"
          },
          "version": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UpdateWorldRoomRequest": {
        "required": [
          "name",
          "description",
          "accessType",
          "version"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "accessType": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}